clean the code

+++ backend/src/main/java/com/controller/BlogController.java
@@ -3,6 +3,8 @@ package com.controller;
 import com.dto.Request.BlogRequest;
 import com.dto.Request.CommentRequestDTO;
 import com.dto.Request.LikeOrSaveBlogRequest;
+import com.dto.Request.PageableBlogsDTO;
+import com.dto.Request.PageableDTO;
 import com.dto.Response.BlogResponseDTO;
 import com.dto.Response.CommentResponseDTO;
 import com.entity.UserDetailsImpl;
@@ -11,7 +13,6 @@ import com.service.Blogs.CommentBlogService;
 import com.service.Blogs.LikeBlogService;
 import com.service.Blogs.SavedService;
 import java.util.List;
-import java.util.Map;
 import java.util.UUID;
 
 import com.util.Response;
@@ -38,6 +39,9 @@ public class BlogController {
     @DeleteMapping("/deleteblog/{blogId}")
     public ResponseEntity<Response<?>> deleteBlog(@PathVariable UUID blogId,
             @AuthenticationPrincipal UserDetailsImpl userDetails) {
+        if (blogId == null) {
+            return ResponseEntity.badRequest().body(new Response<>(false, "blog_id cannot be null"));
+        }
         try {
             UUID user_id = userDetails.getUser().getId();
             Response<?> blogstatus = blogService.deleteOneBlog(user_id, blogId);
@@ -48,29 +52,53 @@ public class BlogController {
     }
 
     @PostMapping("/like_blog")
-    public ResponseEntity<?> likeBlog(@RequestBody LikeOrSaveBlogRequest request,
+    public ResponseEntity<Response<?>> likeBlog(@RequestBody @Valid LikeOrSaveBlogRequest request,
+            BindingResult bindingResult,
             @AuthenticationPrincipal UserDetailsImpl userDetails) {
-        UUID user_id = userDetails.getUser().getId();
-        String msg = this.likeBlogService.likeBlog(user_id, request.getId_blog());
-        return ResponseEntity.ok(Map.of("message", msg));
+        if (bindingResult.hasErrors()) {
+            String errorMsg = bindingResult.getAllErrors().get(0).getDefaultMessage();
+            return ResponseEntity.badRequest().body(new Response<>(false, errorMsg));
+        }
+
+        try {
+            UUID user_id = userDetails.getUser().getId();
+            Response<?> data = this.likeBlogService.likeBlog(user_id, request.getId_blog());
+            return ResponseEntity.status(HttpStatus.CREATED).body(data);
+        } catch (Exception e) {
+            return ResponseEntity.badRequest().body(new Response<>(false, e.getMessage()));
+        }
     }
 
     @PostMapping("/unlike_blog")
-    public ResponseEntity<?> UnLikeBlog(@RequestBody LikeOrSaveBlogRequest request,
+    public ResponseEntity<Response<?>> UnLikeBlog(@RequestBody @Valid LikeOrSaveBlogRequest request,
+            BindingResult bindingResult,
             @AuthenticationPrincipal UserDetailsImpl userDetails) {
 
-        UUID user_id = userDetails.getUser().getId();
-        String msg = this.likeBlogService.unLikedBlog(user_id, request.getId_blog());
+        if (bindingResult.hasErrors()) {
+            String errorMsg = bindingResult.getAllErrors().get(0).getDefaultMessage();
+            return ResponseEntity.badRequest().body(new Response<>(false, errorMsg));
+        }
+        try {
+            UUID user_id = userDetails.getUser().getId();
+            Response<?> data = this.likeBlogService.unLikedBlog(user_id, request.getId_blog());
 
-        return ResponseEntity.ok(Map.of("message", msg));
+            return ResponseEntity.status(HttpStatus.CREATED).body(data);
+        } catch (Exception e) {
+            return ResponseEntity.badRequest().body(new Response<>(false, e.getMessage()));
+        }
     }
 
     @PostMapping("/save_blog")
-    public ResponseEntity<Response<?>> saveBlog(@RequestBody LikeOrSaveBlogRequest request,
+    public ResponseEntity<Response<?>> saveBlog(@RequestBody @Valid LikeOrSaveBlogRequest request,
+            BindingResult bindingResult,
             @AuthenticationPrincipal UserDetailsImpl userDetails) {
 
-        UUID user_id = userDetails.getUser().getId();
+        if (bindingResult.hasErrors()) {
+            String errorMsg = bindingResult.getAllErrors().get(0).getDefaultMessage();
+            return ResponseEntity.badRequest().body(new Response<>(false, errorMsg));
+        }
         try {
+            UUID user_id = userDetails.getUser().getId();
             this.savedService.saveBlog(user_id, request.getId_blog());
             return ResponseEntity.status(HttpStatus.CREATED).body(new Response<>(true, "saved successfully!"));
         } catch (Exception e) {
@@ -80,12 +108,17 @@ public class BlogController {
     }
 
     @PostMapping("/unsave_blog")
-    public ResponseEntity<Response<?>> unsaveBlog(@RequestBody LikeOrSaveBlogRequest request,
+    public ResponseEntity<Response<?>> unsaveBlog(@RequestBody @Valid LikeOrSaveBlogRequest request,
+            BindingResult bindingResult,
             @AuthenticationPrincipal UserDetailsImpl userDetails) {
 
-        UUID user_id = userDetails.getUser().getId();
+        if (bindingResult.hasErrors()) {
+            String errorMsg = bindingResult.getAllErrors().get(0).getDefaultMessage();
+            return ResponseEntity.badRequest().body(new Response<>(false, errorMsg));
+        }
 
         try {
+            UUID user_id = userDetails.getUser().getId();
             this.savedService.unsaveBlog(user_id, request.getId_blog());
             return ResponseEntity.status(HttpStatus.CREATED).body(new Response<>(true, "Unsaved successfully!"));
         } catch (Exception e) {
@@ -96,23 +129,34 @@ public class BlogController {
 
     @GetMapping("/blogs")
     public ResponseEntity<Response<List<BlogResponseDTO>>> getBlogs(
-            @RequestParam int page,
-            @RequestParam int size,
-            @RequestParam String mode,
+            @ModelAttribute @Valid PageableBlogsDTO pageableDTO, BindingResult bindingResult,
             @RequestParam(required = false) String username, @AuthenticationPrincipal UserDetailsImpl userDetails) {
+        if (bindingResult.hasErrors()) {
+            String errorMsg = bindingResult.getAllErrors().get(0).getDefaultMessage();
+            return ResponseEntity.badRequest().body(new Response<>(false, errorMsg));
+        }
 
         try {
             UUID userId = userDetails.getUser().getId();
+            switch (pageableDTO.getMode()) {
+                case HOME:
+                    List<BlogResponseDTO> blogDTOhome = blogService.blogsGetterHome(userId, pageableDTO.getPage(),
+                            pageableDTO.getSize());
+                    return ResponseEntity.ok(new Response<>(true, "getBlogs Sucesfuly", blogDTOhome));
 
-            if ("home".equals(mode)) {
-                List<BlogResponseDTO> blogDTOs = blogService.blogsGetterHome(userId, page, size);
-                return ResponseEntity.ok(new Response<>(true, "getBlogs Sucesfuly", blogDTOs));
-            } else if ("profile".equals(mode)) {
-                List<BlogResponseDTO> blogDTOs = blogService.blogsGetterProfile(userId, page, size, username);
-                return ResponseEntity.ok(new Response<>(true, "getBlogs Sucesfuly", blogDTOs));
-            } else if ("saved".equals(mode)) {
-                List<BlogResponseDTO> blogDTOs = blogService.blogsGetterSaved(userId, page, size);
-                return ResponseEntity.ok(new Response<>(true, "getBlogs Sucesfuly", blogDTOs));
+                case PROFILE:
+                    List<BlogResponseDTO> blogDTOprofile = blogService.blogsGetterProfile(userId, pageableDTO.getPage(),
+                            pageableDTO.getSize(), username);
+                    return ResponseEntity.ok(new Response<>(true, "getBlogs Sucesfuly", blogDTOprofile));
+
+                case SAVED:
+                    List<BlogResponseDTO> blogDTOsaved = blogService.blogsGetterSaved(userId, pageableDTO.getPage(),
+                            pageableDTO.getSize());
+                    return ResponseEntity.ok(new Response<>(true, "getBlogs Sucesfuly", blogDTOsaved));
+
+                default:
+                    break;
+
             }
         } catch (Exception e) {
             return ResponseEntity.badRequest().body(new Response<>(false, e.getMessage()));
@@ -121,34 +165,34 @@ public class BlogController {
     }
 
     @GetMapping("/blog/{id_blog}")
-    public ResponseEntity<?> getBlog(@PathVariable UUID id_blog,
+    public ResponseEntity<Response<?>> getBlog(@PathVariable UUID id_blog,
             @AuthenticationPrincipal UserDetailsImpl userDetails) {
-
         try {
             UUID user_id = userDetails.getUser().getId();
             BlogResponseDTO blogDTOs = blogService.getOneBlog(user_id, id_blog);
-            return ResponseEntity.ok(blogDTOs);
+            return ResponseEntity.ok(new Response<>(true, null, blogDTOs));
         } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Blog not found or error occurred");
+            return ResponseEntity.badRequest().body(new Response<>(false, e.getMessage()));
         }
     }
 
     @GetMapping("/comments/{id_blog}")
     public ResponseEntity<Response<?>> getcomments(@PathVariable UUID id_blog,
-            @RequestParam(defaultValue = "0") int page,
-            @RequestParam(defaultValue = "10") int size, @AuthenticationPrincipal UserDetailsImpl userDetails) {
-        UUID user_id = userDetails.getUser().getId();
-
+            @ModelAttribute @Valid PageableDTO pageableDTO, BindingResult bindingResult,
+            @AuthenticationPrincipal UserDetailsImpl userDetails) {
+        if (bindingResult.hasErrors()) {
+            String errorMsg = bindingResult.getAllErrors().get(0).getDefaultMessage();
+            return ResponseEntity.badRequest().body(new Response<>(false, errorMsg));
+        }
+        
         try {
-            List<CommentResponseDTO> CommentDTO = this.commentBlogService.getTheComment(user_id, page, size, id_blog);
-
+            UUID user_id = userDetails.getUser().getId();
+            List<CommentResponseDTO> CommentDTO = this.commentBlogService.getTheComment(user_id, pageableDTO.getPage(),
+                    pageableDTO.getSize(), id_blog);
             return ResponseEntity.ok(new Response<>(true, "", CommentDTO));
-
         } catch (Exception e) {
-
             return ResponseEntity.badRequest().body(new Response<>(false, e.getMessage()));
         }
-
     }
 
     @PostMapping("/creat-comment")
